refactor(cartpost): 외부 통신용 컨트롤러를 external 패키지로 따로 이동하여 관리
