--- cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/dto/response/CommissionCreateResponse.java
@@ -1,7 +1,0 @@
-package com.example.cartpostservice.commissions.controller.dto.response;
-
-public record CommissionCreateResponse(
-        String commissionsCode
-) {
-
-}
