+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/external/dto/response/CommissionCreateResponse.java
@@ -1,0 +1,7 @@
+package com.example.cartpostservice.commissions.controller.external.dto.response;
+
+public record CommissionCreateResponse(
+        String commissionsCode
+) {
+
+}
