+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/external/dto/response/CommissionUpdateResponse.java
@@ -1,0 +1,7 @@
+package com.example.cartpostservice.commissions.controller.external.dto.response;
+
+public record CommissionUpdateResponse(
+        String commissionsCode
+) {
+
+}
